# src/main/resources/static/api-docs.yaml
openapi: 3.0.1
info:
    title: 'MilliMeeter Server'
    description: 'Just a dating app'
    license:
        name: 'Apache 2.0'
        url: 'http://springdoc.org'
    version: v0.0.1
...
servers:
    -
        url: 'http://localhost:8080'
        description: 'Generated server url'
paths:
    /api/v1/profiles:
        get:
            tags:
                - Profile
            summary: 'Find profiles to swipe'
            operationId: findProfilesToSwipe
            responses:
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                200:
                    description: OK
                    content:
                        application/hal+json:
                            examples:
                                'Found profiles to swipe':
                                    description: 'Found profiles to swipe'
                                'No profiles to swipe found':
                                    description: 'No profiles to swipe found'
            security:
                -
                    bearer-key: []
        put:
            tags:
                - Profile
            summary: 'Update the profile'
            operationId: update
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdatedProfileDto'
                required: true
            responses:
                200:
                    description: OK
                    content:
                        application/hal+json:
                            examples:
                                'Profile updated':
                                    description: 'Profile updated'
                400:
                    description: 'BAD REQUEST'
                    content:
                        application/json:
                            examples:
                                'Bad request':
                                    description: 'Bad request'
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                        application/json:
                            examples:
                                'Not valid update request':
                                    description: 'Not valid update request'
            security:
                -
                    bearer-key: []
        post:
            tags:
                - Profile
            summary: 'Create the profile'
            operationId: create_1
            requestBody:
                content:
                    application/json:
                        schema:
                            required:
                                - body
                                - photos
                            type: object
                            properties:
                                body:
                                    $ref: '#/components/schemas/RegistrationDto'
                                photos:
                                    type: array
                                    items:
                                        type: string
                                        format: binary
            responses:
                409:
                    description: CONFLICT
                    content:
                        text/plain:
                            examples:
                                'Profile already exists':
                                    description: 'Profile already exists'
                                    value: 'Profile already exists'
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Profile not created, too many photos uploaded':
                                    description: 'Profile not created, too many photos uploaded'
                                    value: 'Limit of 5 photos exceeded'
                201:
                    description: CREATED
                    content:
                        application/hal+json:
                            examples:
                                'Profile created':
                                    description: 'Profile created'
            security:
                -
                    bearer-key: []
        delete:
            tags:
                - Profile
            summary: 'Delete my profile'
            operationId: deleteProfile
            responses:
                204:
                    description: 'NO CONTENT'
                    content:
                        text/plain:
                            examples:
                                'Profile deleted':
                                    description: 'Profile deleted'
            security:
                -
                    bearer-key: []
    /api/v1/profiles/photos:
        put:
            tags:
                - Profile
            summary: 'Upload the photo'
            operationId: uploadPhoto
            parameters:
                -
                    name: index
                    in: query
                    required: true
                    schema:
                        maximum: 4
                        minimum: -1
                        type: integer
                        format: int32
            requestBody:
                content:
                    application/json:
                        schema:
                            required:
                                - photos
                            type: object
                            properties:
                                photos:
                                    type: string
                                    format: binary
            responses:
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Photo not uploaded, limit already reached':
                                    description: 'Photo not uploaded, limit already reached'
                                    value: 'Limit of 5 photos reached'
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                        application/json:
                            examples:
                                'Not valid photo upload request':
                                    description: 'Not valid photo upload request'
                200:
                    description: OK
                    content:
                        application/hal+json:
                            examples:
                                'Uploaded photo as last':
                                    description: 'Uploaded photo as last'
                                'Uploaded photo at certain index':
                                    description: 'Uploaded photo at certain index'
                400:
                    description: 'BAD REQUEST'
                    content:
                        text/plain:
                            examples:
                                'Bad request':
                                    description: 'Bad request'
            security:
                -
                    bearer-key: []
    '/api/v1/profiles/photos/{index}':
        put:
            tags:
                - Profile
            summary: 'Delete the photo by index'
            operationId: deletePhoto
            parameters:
                -
                    name: index
                    in: path
                    required: true
                    schema:
                        maximum: 4
                        minimum: -1
                        type: integer
                        format: int32
            responses:
                400:
                    description: 'BAD REQUEST'
                    content:
                        text/plain:
                            examples:
                                'Bad request':
                                    description: 'Bad request'
                200:
                    description: OK
                    content:
                        application/json:
                            examples:
                                'Photo deleted':
                                    description: 'Photo deleted'
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                        application/json:
                            examples:
                                'Not valid delete photo request':
                                    description: 'Not valid delete photo request'
            security:
                -
                    bearer-key: []
    /api/v1/profiles/location:
        put:
            tags:
                - Profile
            summary: 'Update the location'
            operationId: updateLocation
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LocationDto'
                required: true
            responses:
                400:
                    description: 'BAD REQUEST'
                    content:
                        application/json:
                            examples:
                                'Bad request':
                                    description: 'Bad request'
                200:
                    description: OK
                    content:
                        application/hal+json:
                            examples:
                                'Location updated':
                                    description: 'Location updated'
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        application/json:
                            examples:
                                'Not valid update request':
                                    description: 'Not valid update request'
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
            security:
                -
                    bearer-key: []
...
    /api/v1/profiles/swipesLeft:
        get:
            tags:
                - Profile
            summary: 'Get swipes left count'
            operationId: getSwipesLeftCount
            responses:
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Profile not exists':
                                    description: 'Profile not exists'
                                    value: 'Profile not exists'
                200:
                    description: OK
                    content:
                        text/plain:
                            examples:
                                'Got swipes left number':
                                    description: 'Got swipes left number'
            security:
                -
                    bearer-key: []
    /api/v1/profiles/me:
        get:
            tags:
                - Profile
            summary: 'Get my profile'
            operationId: getMyProfile
            responses:
                422:
                    description: 'UNPROCESSABLE ENTITY'
                    content:
                        text/plain:
                            examples:
                                'Profile not exist':
                                    description: 'Profile not exist'
                                    value: 'Profile not exists'
                200:
                    description: OK
                    content:
                        application/hal+json:
                            examples:
                                'Got my profile':
                                    description: 'Got my profile'
            security:
                -
                    bearer-key: []
...
components:
    schemas:
        UpdatedProfileDto:
            required:
                - ageRangeMaximum
                - ageRangeMinimum
                - lastLatitude
                - lastLongitude
                - lookingFor
                - mySong
                - searchDistance
            type: object
            properties:
                description:
                    maxLength: 500
                    minLength: 0
                    type: string
                mySong:
                    pattern: '[a-zA-Z0-9]*'
                    type: string
                lastLatitude:
                    maximum: 90
                    minimum: -90
                    type: number
                    format: double
                lastLongitude:
                    maximum: 90
                    minimum: -90
                    type: number
                    format: double
                lookingFor:
                    pattern: MEN|WOMEN|BOTH
                    type: string
                searchDistance:
                    maximum: 100
                    minimum: 1
                    type: integer
                    format: int32
                ageRangeMinimum:
                    maximum: 100
                    minimum: 18
                    type: integer
                    format: int32
                ageRangeMaximum:
                    maximum: 100
                    minimum: 18
                    type: integer
                    format: int32
        LocationDto:
            required:
                - lastLatitude
                - lastLongitude
            type: object
            properties:
                lastLatitude:
                    maximum: 90
                    minimum: -90
                    type: number
                    format: double
                lastLongitude:
                    maximum: 90
                    minimum: -90
                    type: number
                    format: double
        RegistrationDto:
            required:
                - ageRangeMaximum
                - ageRangeMinimum
                - dateOfBirth
                - firstName
                - gender
                - lastLatitude
                - lastLongitude
                - lookingFor
                - mySong
                - searchDistance
            type: object
            properties:
                firstName:
                    pattern: '[A-Z][a-z]{1,19}'
                    type: string
                dateOfBirth:
                    type: string
                    format: date
                gender:
                    pattern: MAN|WOMAN
                    type: string
                description:
                    maxLength: 500
                    minLength: 0
                    type: string
                mySong:
                    pattern: '[a-zA-Z0-9]*'
                    type: string
                lastLatitude:
                    maximum: 90
                    minimum: -90
                    type: number
                    format: double
                lastLongitude:
                    maximum: 90
                    minimum: -90
                    type: number
                    format: double
                lookingFor:
                    pattern: MEN|WOMEN|BOTH
                    type: string
                searchDistance:
                    maximum: 100
                    minimum: 1
                    type: integer
                    format: int32
                ageRangeMinimum:
                    maximum: 100
                    minimum: 18
                    type: integer
                    format: int32
                ageRangeMaximum:
                    maximum: 100
                    minimum: 18
                    type: integer
                    format: int32
...
    securitySchemes:
        bearer-key:
            type: http
            scheme: bearer
            bearerFormat: JWT